Don't follow 302 redirects in Sentry API client
It seems that the Sentry API has changed such that it now returns
302 responses with a Location header when a resource is moved or
orphaned.

When the client tries to follow this 302 redirect, the original
HTTP method and body are no longer preserved, instead falling back
to a GET request instead which might break how the response body
is unmarshalled into our data structures.

To work around this, we disable following of HTTP redirects, instead
choosing to surface the error that the resource has been moved.